+++ clients/TypeScript/packages/client/src/StateQuery/queries/delegationsAndRewards.ts
@@ -1,6 +1,8 @@
 import {
-  DelegationsAndRewards1,
-  EraMismatch, Hash16,
+  DelegationsAndRewardsByAccounts,
+  DelegationsAndRewards,
+  EraMismatch,
+  Hash16,
   Ogmios
 } from '@cardano-ogmios/schema'
 import { EraMismatchError, QueryUnavailableInCurrentEraError, UnknownResultError } from '../../errors'
@@ -10,16 +12,16 @@ import { Query } from '../Query'
 const isEraMismatch = (result: Ogmios['QueryResponse[delegationsAndRewards]']['result']): result is EraMismatch =>
   (result as EraMismatch).eraMismatch !== undefined
 
-const isDelegationsAndRewards = (result: Ogmios['QueryResponse[delegationsAndRewards]']['result']): result is DelegationsAndRewards1 => {
-  const sample = Object.entries(result as DelegationsAndRewards1)[0]
+const isDelegationsAndRewardsByAccounts = (result: Ogmios['QueryResponse[delegationsAndRewards]']['result']): result is DelegationsAndRewardsByAccounts => {
+  const sample = Object.entries(result as DelegationsAndRewards)[0]
   return typeof sample[0] === 'string' && (sample[1].delegate !== undefined || sample[1].rewards !== undefined)
 }
 
-export const delegationsAndRewards = (stakeKeyHashes: Hash16[], context?: InteractionContext): Promise<DelegationsAndRewards1> =>
+export const delegationsAndRewards = (stakeKeyHashes: Hash16[], context?: InteractionContext): Promise<DelegationsAndRewardsByAccounts> =>
   Query<
     Ogmios['Query'],
     Ogmios['QueryResponse[delegationsAndRewards]'],
-    DelegationsAndRewards1
+    DelegationsAndRewardsByAccounts
   >({
     methodName: 'Query',
     args: {
@@ -33,7 +35,7 @@ export const delegationsAndRewards = (stakeKeyHashes: Hash16[], context?: Intera
         const { eraMismatch } = response.result
         const { ledgerEra, queryEra } = eraMismatch
         return reject(new EraMismatchError(queryEra, ledgerEra))
-      } else if (isDelegationsAndRewards(response.result)) {
+      } else if (isDelegationsAndRewardsByAccounts(response.result)) {
         return resolve(response.result)
       } else {
         return reject(new UnknownResultError(response.result))